Make ScoreManager have a static instance and change score from float to ulong

+++ Assets/Scripts/UI/Leaderboard/Score.cs
@@ -6,9 +6,9 @@ using System;
 [Serializable]
 public class Score {
     public string name;
-    public float score;
+    public ulong score;
 
-    public Score(string name, float score) {
+    public Score(string name, ulong score) {
         this.name = name;
         this.score = score;
     }
